+++ src/components/pages/EdcConnector/index.tsx
@@ -52,6 +52,11 @@ import { SuccessErrorType } from 'features/admin/appuserApiSlice'
 import { ManagedConnectorTableColumns } from './edcManagedConnectorTableColumns'
 import { OwnConnectorTableColumns } from './edcOwnConnectorTableColumns'
 import ConfigurationDetailsOverlay from './ConfigurationDetailsOverlay'
+import {
+  ServiceAccountType,
+  useAddServiceAccountMutation,
+  useFetchServiceAccountRolesQuery,
+} from 'features/admin/serviceApiSlice'
 
 const EdcConnector = () => {
   const { t } = useTranslation()
@@ -98,6 +103,20 @@ const EdcConnector = () => {
     viewConfigurationDetailsOverlayOpen,
     setViewConfigurationDetailsOverlayOpen,
   ] = useState<boolean>(false)
+  const [addServiceAccount] = useAddServiceAccountMutation()
+  const roles = useFetchServiceAccountRolesQuery().data
+  const [role, setRole] = useState('')
+  const [newUserLoading, setNewUserLoading] = useState<boolean>(false)
+  const [newUserSuccess, setNewUserSuccess] = useState<boolean>(false)
+  const [serviceAccId, setServiceAccId] = useState('')
+
+  useEffect(() => {
+    if (roles && roles.length > 0)
+      setRole(
+        roles?.filter((i) => i.roleName === 'Identity Wallet Management')[0]
+          .roleId
+      )
+  }, [roles])
 
   const onStepChange = () => {
     setAddConnectorOverlayCurrentStep(0)
@@ -148,6 +167,10 @@ const EdcConnector = () => {
     body.append('Status', ConnectorStatusType.PENDING)
     setLoading(true)
     if (selectedService.type === ConnectType.COMPANY_CONNECTOR) {
+      body.append(
+        'TechnicalUserId',
+        serviceAccId === '' ? data.ConnectorTechnicalUser : serviceAccId
+      )
       await createConnector(body)
         .unwrap()
         .then(() => {
@@ -166,9 +189,33 @@ const EdcConnector = () => {
           showOverlay(true)
         })
         .catch(() => {
-          showOverlay(false)
+          !setNewUserSuccess && showOverlay(false)
         })
     }
+  }
+
+  const onSubmitClick = async (data: FormFieldsType) => {
+    const body = new FormData()
+    body.append('name', data.TechnicalUserName)
+    body.append('description', data.TechnicalUserDescription)
+    body.append('authenticationType', ServiceAccountType.SECRET)
+    body.append('roleIds', role)
+    setNewUserLoading(true)
+    await addServiceAccount({
+      name: data.TechnicalUserName,
+      description: data.TechnicalUserDescription,
+      authenticationType: ServiceAccountType.SECRET,
+      roleIds: [role],
+    })
+      .unwrap()
+      .then((res) => {
+        setServiceAccId(res.serviceAccountId)
+        setNewUserSuccess(true)
+      })
+      .catch(() => {
+        showOverlay(false)
+      })
+    setNewUserLoading(false)
   }
 
   const showOverlay = (result: boolean) => {
@@ -272,7 +319,10 @@ const EdcConnector = () => {
         connectorStep={addConnectorOverlayCurrentStep}
         handleConfirmClick={onConfirmClick}
         onFormConfirmClick={(data) => void onFormSubmit(data)}
+        onSubmitClick={(data) => void onSubmitClick(data)}
         loading={loading}
+        newUserLoading={newUserLoading}
+        newUserSuccess={newUserSuccess}
         onStepChange={onStepChange}
       />
       <ConfigurationDetailsOverlay